requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.33.0
backoff==2.2.1
blinker==1.9.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.3
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
curlify==2.2.1
Eel==0.16.0
exceptiongroup==1.2.0
facebook-business==18.0.5
facebookads==2.11.4
Flask==3.1.0
Flask-Cors==5.0.0
flask-ngrok==0.0.25
frozenlist==1.4.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
greenlet==2.0.2
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.1.4
pefile==2023.2.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==23.12.11
pycparser==2.21
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.1.3
whichcraft==0.6.1
wsproto==1.2.0
yarl==1.9.4
zope.event==4.6
zope.interface==6.0

Flask==3.1.0
flask-cors==5.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.29.0

